Release 1.0.0 (#10)
* Update gradle wrapper

* Add request builder

* WIP

* Fix debug login

* Add DebugActivity

* Update .gitignore

* Update .gitignore

* Added balance fragment

Fragment for displaying the balance on multiple fragments.

* Shared data between fragments

Added a ViewModel for storing data between fragment lifcycles.

* Appendix expantion.

Made the appendix funtion scalable. Added a placeholder for medicine worth.

* Name changes and desctiption.

Changed the name of medicines to medicine.
Added some descriptions to the functions in SharedGameDataViewmodel.

* Mediaction

Added some stuff for medication and use it in dataviewmodel

* strings to us_en

* Add account fragment

* Add balance to account view

* Fix nullpointer in FadingTextView

* Medicine store and user medicine

not funtional yet but functions have been added.

* Medication and traits

Functions for user and medication have been updated.

* localUser in funcitons

* Working

Able to get Medicines from back-end.

* W.I.P

* W.I.P

* W.I.P

* W.I.P

* Click animations

When you click a indicator with the value gets show on that location.

* added medicine fragment

* Intermedite progress

need to push to rebase for medicine store.

* Medication page

In progress

* temp

* card errors

* tiers

* Update Medication.java

* W.I.P

* W.I.P

* W.I.P

* W.I.P

* FriendStore

Added the friend stuff to user store

* appedix fixed

* Update SharedGameDataViewModel.java

* removed medicine fragment

* Add loading screen

* Fix backstack bug

* Add event handler

* Add disease fragment

* Add database post

* Friends

gathers friends and displays them in the fragment. TODO add end respond to friend requests.

* Dialog and firebase Messaging

* first massages

* friends :)

* tabs!

* codes

* stuff

* Update Registrar.java

* Update UserStore.java

* Fix delete friend

* Fix event interval

* Update launcher icon

Co-authored-by: Arjan <arjan@koens.biz>
Co-authored-by: Akoens <43842651+Akoens@users.noreply.github.com>